Move CORS config to app and add tests
